Create spaceeater.py
In Space Eater, the player Turtle tries to eat as many red targets as possible, which move in random directions around the boundary box. Every time the Turtle or a target hits the boundary, it bounces off in the opposite direction (i.e. turns 180 degrees). The Turtle is continuously moving forward by itself, and the left and right arrow keys control its direction.